+++ R/bounds.r
@@ -302,7 +302,9 @@ squish_infinite <- oob_squish_infinite
 #' @param zero_width distance to use if range has zero width
 #' @export
 expand_range <- function(range, mul = 0, add = 0, zero_width = 1) {
-  if (is.null(range)) return()
+  if (is.null(range)) {
+    return()
+  }
 
   width <- if (zero_range(range)) zero_width else diff(range)
 
@@ -319,26 +321,26 @@ expand_range <- function(range, mul = 0, add = 0, zero_width = 1) {
 #'
 #' @examples
 #' eps <- .Machine$double.eps
-#' zero_range(c(1, 1 + eps))       # TRUE
-#' zero_range(c(1, 1 + 99 * eps))  # TRUE
-#' zero_range(c(1, 1 + 1001 * eps)) # FALSE - Crossed the tol threshold
-#' zero_range(c(1, 1 + 2 * eps), tol = eps) # FALSE - Changed tol
+#' zero_range(c(1, 1 + eps))
+#' zero_range(c(1, 1 + 99 * eps))
+#' zero_range(c(1, 1 + 1001 * eps))
+#' zero_range(c(1, 1 + 2 * eps), tol = eps)
 #'
 #' # Scaling up or down all the values has no effect since the values
 #' # are rescaled to 1 before checking against tol
-#' zero_range(100000 * c(1, 1 + eps))        # TRUE
-#' zero_range(100000 * c(1, 1 + 1001 * eps))  # FALSE
-#' zero_range(.00001 * c(1, 1 + eps))        # TRUE
-#' zero_range(.00001 * c(1, 1 + 1001 * eps))  # FALSE
+#' zero_range(100000 * c(1, 1 + eps))
+#' zero_range(100000 * c(1, 1 + 1001 * eps))
+#' zero_range(.00001 * c(1, 1 + eps))
+#' zero_range(.00001 * c(1, 1 + 1001 * eps))
 #'
 #' # NA values
-#' zero_range(c(1, NA))   # NA
-#' zero_range(c(1, NaN))  # NA
+#' zero_range(c(1, NA)) # NA
+#' zero_range(c(1, NaN)) # NA
 #'
 #' # Infinite values
-#' zero_range(c(1, Inf))     # FALSE
-#' zero_range(c(-Inf, Inf))  # FALSE
-#' zero_range(c(Inf, Inf))   # TRUE
+#' zero_range(c(1, Inf)) # FALSE
+#' zero_range(c(-Inf, Inf)) # FALSE
+#' zero_range(c(Inf, Inf)) # TRUE
 #'
 #' @export
 #' @param x numeric range: vector of length 2
@@ -346,23 +348,33 @@ expand_range <- function(range, mul = 0, add = 0, zero_width = 1) {
 #' @return logical `TRUE` if the relative difference of the endpoints of
 #' the range are not distinguishable from 0.
 zero_range <- function(x, tol = 1000 * .Machine$double.eps) {
-  if (length(x) == 1) return(TRUE)
+  if (length(x) == 1) {
+    return(TRUE)
+  }
   if (length(x) != 2) stop("x must be length 1 or 2")
-  if (any(is.na(x))) return(NA)
+  if (any(is.na(x))) {
+    return(NA)
+  }
 
   # Special case: if they are equal as determined by ==, then there
   # is zero range. Also handles (Inf, Inf) and (-Inf, -Inf)
-  if (x[1] == x[2]) return(TRUE)
+  if (x[1] == x[2]) {
+    return(TRUE)
+  }
 
   # If we reach this, then x must be (-Inf, Inf) or (Inf, -Inf)
-  if (all(is.infinite(x))) return(FALSE)
+  if (all(is.infinite(x))) {
+    return(FALSE)
+  }
 
   # Take the smaller (in magnitude) value of x, and use it as the scaling
   # factor.
   m <- min(abs(x))
 
   # If we get here, then exactly one of the x's is 0. Return FALSE
-  if (m == 0) return(FALSE)
+  if (m == 0) {
+    return(FALSE)
+  }
 
   # If x[1] - x[2] (scaled to 1) is smaller than tol, then return
   # TRUE; otherwise return FALSE