+++ R/breaks-retired.R
@@ -1,6 +1,6 @@
 #' Regularly spaced dates
 #'
-#' \Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
+#' `r lifecycle::badge('superseded')`
 #' Use `breaks_width()` instead.
 #'
 #' @param width an interval specification, one of "sec", "min", "hour",
@@ -15,7 +15,7 @@ date_breaks <- function(width = "1 month") {
 
 #' Pretty breaks on transformed scale
 #'
-#' \Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
+#' `r lifecycle::badge('superseded')`
 #' These often do not produce very attractive breaks.
 #'
 #' @param trans function of single variable, `x`, that given a numeric
@@ -26,8 +26,8 @@ date_breaks <- function(width = "1 month") {
 #' @keywords internal
 #' @export
 #' @examples
-#' trans_breaks("log10", function(x) 10 ^ x)(c(1, 1e6))
-#' trans_breaks("sqrt", function(x) x ^ 2)(c(1, 100))
+#' trans_breaks("log10", function(x) 10^x)(c(1, 1e6))
+#' trans_breaks("sqrt", function(x) x^2)(c(1, 100))
 #' trans_breaks(function(x) 1 / x, function(x) 1 / x)(c(1, 100))
 #' trans_breaks(function(x) -x, function(x) -x)(c(1, 100))
 trans_breaks <- function(trans, inv, n = 5, ...) {
@@ -42,7 +42,7 @@ trans_breaks <- function(trans, inv, n = 5, ...) {
 
 #' Compute breaks for continuous scale
 #'
-#' \Sexpr[results=rd, stage=render]{lifecycle::badge("retired")}
+#' `r lifecycle::badge('superseded')`
 #' This function wraps up the components needed to go from a continuous range
 #' to a set of breaks and labels suitable for display on axes or legends.
 #'
@@ -73,8 +73,10 @@ trans_breaks <- function(trans, inv, n = 5, ...) {
 #' # You can also specify them manually:
 #' cbreaks(c(0, 100), breaks = c(15, 20, 80))
 #' cbreaks(c(0, 100), breaks = c(15, 20, 80), labels = c(1.5, 2.0, 8.0))
-#' cbreaks(c(0, 100), breaks = c(15, 20, 80),
-#'   labels = expression(alpha, beta, gamma))
+#' cbreaks(c(0, 100),
+#'   breaks = c(15, 20, 80),
+#'   labels = expression(alpha, beta, gamma)
+#' )
 cbreaks <- function(range, breaks = extended_breaks(), labels = scientific_format()) {
   if (zero_range(range)) {
     return(list(breaks = range[1], labels = format(range[1])))