+++ R/label-expression.R
@@ -5,7 +5,7 @@
 #' with each string.
 #'
 #' @section Old interface:
-#' `parse_format()` and `math_format()` was retired; please use
+#' `parse_format()` and `math_format()` was superseded; please use
 #' `label_parse()` and `label_math()` instead.
 #' @inherit number_format return
 #' @seealso [plotmath] for the details of mathematical formatting in R.