+++ NEWS.md
@@ -1,5 +1,16 @@
 # scales (development version)
 
+* `hue_pal()` respects `h.start` once again (#288).
+
+* `col_quantile()` no longer errors if data is sufficiently skewed that we
+  can't generate the requested number of unique colours (#294).
+
+* `breaks_width()` now supports units like `"3 months"` in the `offset`
+  argument.
+
+* `label_number_auto()` correctly formats single numbers that are greater than
+  1e+06 without an error (@karawoo, #321)
+
 * `manual_pal()` now always returns an unnamed colour vector, which is easy to
   use with `ggplot2::discrete_scale()` (@yutannihilation, #284).
 
@@ -24,6 +35,9 @@
   
 * `label_bytes()` now correctly accounts for the `scale` argument when choosing
   auto units (@davidchall, #235).
+  
+* Internal `precision()`, used when `accuracy = NULL`, now avoids displaying
+  unnecessary digits (@davidchall, #304).
 
 * New `label_log()` displays the base and a superscript exponent, for use with
   logarithmic axes (@davidchall, #312).